fix scrolling section
