+++ src/components/layout/ScrollAnimationStepContent.tsx
