+++ src/locales/sv/translation.json
@@ -114,6 +114,38 @@
         }
       }
     },
+    "scrollSteps": {
+      "mission": {
+        "badge": "Transparens för klimatåtgärder",
+        "heading": "Vårt uppdrag",
+        "paragraph": "Klimatkollen finns för att göra klimatdata tillgänglig för alla. Vi tror att transparens driver ansvar, och ansvar driver verklig klimatåtgärd. Genom att sätta utsläppsdata i medborgarnas händer ger vi samhällen möjlighet att hålla organisationer ansvariga för sina miljöåtaganden."
+      },
+      "transparency": {
+        "badge": "Ljuset är det bästa desinfektionsmedlet",
+        "heading": "Varför datatransparens\när viktigt",
+        "paragraph": "När utsläppsdata döljs bakom företagsrapporter och betalda databaser stagnerar framstegen. Transparent, tillgänglig klimatdata skapar marknadstryck för bättre prestanda. Företag och kommuner som vet att deras data är offentlig arbetar hårdare för att förbättra sina rankingar och uppfylla sina klimatåtaganden."
+      },
+      "methodology": {
+        "badge": "Vetenskapsbaserade rankingar",
+        "heading": "Vår metodik",
+        "paragraph": "Vi använder utsläppsdata från officiella rapporteringskällor för att skapa rättvisa, jämförbara rankingar. Våra mätvärden fokuserar på faktiska utsläppsminskningar, inte bara löften eller mål. Varje ranking baseras på verkliga prestationsdata, vilket säkerställer att klimatledare erkänns och klimatfördröjare hålls ansvariga."
+      },
+      "impact": {
+        "badge": "Från data till handling",
+        "heading": "Driva verklig påverkan",
+        "paragraph": "Vår plattform har redan hjälpt till att identifiera klimatledare och de som halkar efter i hela Sverige. Genom att ranka organisationer på deras faktiska utsläppsprestationer skapar vi hälsosam konkurrens som påskyndar klimatframsteg. Datan ljuger inte—och det borde inte klimatåtaganden heller."
+      },
+      "engagement": {
+        "badge": "Varje röst räknas",
+        "heading": "Medborgaråtgärder &\nengagemang",
+        "paragraph": "Klimatåtgärder är inte bara för regeringar och företag—det börjar med informerade medborgare. När du kan se hur din lokala kommun eller favoritföretag presterar på klimatmätvärden kan du fatta bättre beslut som konsument, väljare och samhällsmedlem. Ditt engagemang driver efterfrågan på bättre klimatprestanda."
+      },
+      "pathForward": {
+        "badge": "Bygga ett klimatmedvetet samhälle",
+        "heading": "Vägen framåt",
+        "paragraph": "Föreställ dig en värld där varje organisations klimatprestanda är lika synlig som deras finansiella resultat. Där medborgare enkelt kan jämföra miljöpåverkan av sina val. Där transparens driver en tävling mot toppen i klimatåtgärder. Det är framtiden vi bygger, en datapunkt i taget."
+      }
+    },
     "didYouKnow": {
       "title": "Visste du att?",
       "subtitle": "Intressanta insikter från Sveriges klimatdata - {{action}}!",
